[eve7] move main GED functionality to correspondent controller
